% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{refAnnots}
\alias{refAnnots}
\title{Reference Annotation}
\format{
a data.frame with 15792 rows and 2 variables:
\itemize{
   \item{Source}{the name of the source (fcs) file}
   \item{Cells}{the cell type for the corresponding row in \code{\link{refPhenoMat}}
     or \code{\link{refFuncMat}}}
}
}
\source{
\url{https://reports.cytobank.org/105/v2/}
}
\usage{
refAnnots
}
\description{
A \code{data.frame} containing the source file and cell type for every cell in
\code{\link{refPhenoMat}} and \code{\link{refFuncMat}}. See
\href{https://www.nature.com/articles/nbt.2317/}{Bodenmiller et al 2012}
for details.
}
\keyword{datasets}
