% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refFuncMat}
\alias{refFuncMat}
\title{Reference Matrix (Functional)}
\format{a matrix with 15792 rows and 14 variables:
\itemize{
   \item{ pStat1 }
   \item{ pSlp76 }
   \item{ pBtk }
   \item{ pPlcg2 }
   \item{ pErk }
   \item{ pLat }
   \item{ pS6 }
   \item{ pNFkB }
   \item{ pp38 }
   \item{ pStat5 }
   \item{ pAkt }
   \item{ pSHP2 }
   \item{ pZap70 }
   \item{ pStat3 }
}}
\source{
\url{http://reports.cytobank.org/105/v2}
}
\usage{
refFuncMat
}
\description{
a \code{matrix} containing the signal intensity for functional markers measured using CyTOF
on different cell types. See
\href{http://www.nature.com/nbt/journal/v30/n9/full/nbt.2317.html}{Bodenmiller et al 2012}
for details.
}
\keyword{datasets}

