% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{refPhenoMat}
\alias{refPhenoMat}
\title{Reference Matrix (Phenotypic)}
\format{
a matrix with 15792 rows and 9 variables:
\itemize{
   \item{ CD20 }
   \item{ IgM }
   \item{ CD4 }
   \item{ CD33 }
   \item{ HLA-DR }
   \item{ CD14 }
   \item{ CD7 }
   \item{ CD3 }
   \item{ CD123 }
}
}
\source{
\url{https://reports.cytobank.org/105/v2/}
}
\usage{
refPhenoMat
}
\description{
a \code{matrix} containing the signal intensity for phenotypic markers measured using CyTOF
on different cell types. See
\href{https://www.nature.com/articles/nbt.2317/}{Bodenmiller et al 2012}
for details.
}
\keyword{datasets}
