% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/untreated.R
\docType{data}
\name{untreatedMat}
\alias{untreatedMat}
\title{Untreated Samples with 4 Different Treatments}
\format{a matrix with 15792 rows and 23 variables, 9 phenotypic:
\itemize{
   \item{ CD20 }
   \item{ IgM }
   \item{ CD4 }
   \item{ CD33 }
   \item{ HLA-DR }
   \item{ CD14 }
   \item{ CD7 }
   \item{ CD3 }
   \item{ CD123 }
}
and 14 functional:
\itemize{
   \item{ pStat1 }
   \item{ pSlp76 }
   \item{ pBtk }
   \item{ pPlcg2 }
   \item{ pErk }
   \item{ pLat }
   \item{ pS6 }
   \item{ pNFkB }
   \item{ pp38 }
   \item{ pStat5 }
   \item{ pAkt }
   \item{ pSHP2 }
   \item{ pZap70 }
   \item{ pStat3 }
}}
\source{
\url{http://reports.cytobank.org/105/v2}
}
\usage{
untreatedMat
}
\description{
a \code{matrix} containing the signal intensity for phenotypic and functional markers
measured using CyTOF on different cell types, after stimulation with BCR/FcR-XL, PMA/Ionomycin
and vanadate. See
\href{http://www.nature.com/nbt/journal/v30/n9/full/nbt.2317.html}{Bodenmiller et al 2012}
for details.
}
\keyword{datasets}

