% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{refFuncMat}
\alias{refFuncMat}
\title{Reference Matrix (Functional)}
\format{
a matrix with 15792 rows and 14 variables:
\itemize{
   \item{ pStat1 }
   \item{ pSlp76 }
   \item{ pBtk }
   \item{ pPlcg2 }
   \item{ pErk }
   \item{ pLat }
   \item{ pS6 }
   \item{ pNFkB }
   \item{ pp38 }
   \item{ pStat5 }
   \item{ pAkt }
   \item{ pSHP2 }
   \item{ pZap70 }
   \item{ pStat3 }
}
}
\source{
\url{https://reports.cytobank.org/105/v2/}
}
\usage{
refFuncMat
}
\description{
a \code{matrix} containing the signal intensity for functional markers measured using CyTOF
on different cell types. See
\href{https://www.nature.com/articles/nbt.2317/}{Bodenmiller et al 2012}
for details.
}
\keyword{datasets}
